Split wordpress-develop tests into wpunit and integration
