+++ tests/integration.suite.yml
@@ -18,5 +18,5 @@ modules:
             domain: "%TEST_SITE_WP_DOMAIN%"
             adminEmail: "%TEST_SITE_ADMIN_EMAIL%"
             title: "plugin-slug"
-            plugins: []
-            activatePlugins: []
+            plugins: ['plugin-slug/plugin-slug.php']
+            activatePlugins: ['plugin-slug/plugin-slug.php']
