+++ tests/integration/class-plugin-slug-develop-Test.php
