+++ tests/wpunit/includes/class-i18n-test.php
@@ -16,26 +16,6 @@ namespace Plugin_Package_Name\includes;
 class Plugin_Package_Name_I18n_Test extends \Codeception\TestCase\WPTestCase {
 
 	/**
-	 * AFAICT, this will fail until a translation has been added.
-	 *
-	 * @see load_plugin_textdomain()
-	 * @see https://gist.github.com/GaryJones/c8259da3a4501fd0648f19beddce0249
-	 */
-	public function test_load_plugin_textdomain() {
-
-		$this->markTestSkipped( 'Needs one translation before test might pass.' );
-
-		global $plugin_root_dir;
-
-		$this->assertTrue( file_exists( $plugin_root_dir . '/languages/' ), '/languages/ folder does not exist.' );
-
-		// Seems to fail because there are no translations to load.
-		$this->assertTrue( is_textdomain_loaded( 'plugin-slug' ), 'i18n text domain not loaded.' );
-
-	}
-
-
-	/**
 	 * Checks if the filter run by WordPress in the load_plugin_textdomain() function is called.
 	 *
 	 * @see load_plugin_textdomain()
